+++ packages/uring_api/src/_uring_api/uring_api_core.h
@@ -94,8 +94,7 @@ UringApiCompletion *cqe_get_completion(UringApiRing *self, struct io_uring_cqe *
 unsigned int ring_sq_entries(UringApiRing *self);
 unsigned int ring_cq_entries(UringApiRing *self);
 
-int parse_entries_flags(PyObject *args, PyObject *kwargs, unsigned int default_entries, unsigned int *entries,
-                        unsigned int *flags);
+int parse_entries_flags(URING_API_PARSE_ARGS, unsigned int default_entries, unsigned int *entries, unsigned int *flags);
 int parse_numeric_sockaddr(int fd, PyObject *address, struct sockaddr_storage *storage, socklen_t *addrlen);
 void ring_pending_inc(UringApiRing *self);
 void ring_pending_dec(UringApiRing *self);